Extract CityGML surface mesh tessellator
